# swagger.yaml
openapi: 3.0.0
info:
  title: MEN REST API
  description: MongoDB ExpressJS NodeJS REST API
  version: 1.0.0
servers:
  - url: http://localhost:4000/api/
    description: Localhost development server 
components:
  securitySchemes:
    ApiKeyAuth:        # arbitrary name for the security scheme
      type: apiKey
      in: header       # can be "header", "query" or "cookie"
      name: auth-token

  schemas:
    Movie:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        year:
          type: number
        budget:
          type: number
    
    User:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        date:
          type: string

paths:

  /user/register:
    post:
      summary: Register a new User
      description: Register a new User in the database
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"

      responses:
        201:
          description: User created succesfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  _id:
                    type: string

  /user/login:
    post:
      summary: Login a User
      description: Login a User in order to get JWT token (auth-token)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      
      responses:
        200:
          description: User logged in succesfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  data:
                    type: string


  /Movies:
    get:
      summary: Retrieves a list of Movies
      description: Retrieves a list of Movies as JSON objects.
      responses:
        200:
          description: A list of Movie JSON objects in an array.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Movie"
    
    post:
      summary: Create a new Movie
      description: Create a new Movie
      security:
        - ApiKeyAuth: []      
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Movie"

      responses:
        201:
          description: Movie created succesfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Movie"
  
  /Movies/{id}:
    get:
      summary: Specific Movie
      description: Retrieves a specific Movie based on it id.
      parameters:
        - in: path
          name: id
          required: true
          description: MongoDB id
          schema:
            type: string
      responses:
        200:
          description: A Movie in the format of a JSON object.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Movie"
   
    put:
      summary: Updates a specific Movie
      description: Updates a specific Movie based on it id
      security:
        - ApiKeyAuth: []            
      parameters:
        - in: path
          name: id
          required: true
          description: MongoDB id
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Movie"

      responses:
        201:
          description: Movie updated succesfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Movie"

    delete:
      summary: Deletes a specific Movie
      description: Deletes a specific Movie based on it id
      security:
        - ApiKeyAuth: []            
      parameters:
        - in: path
          name: id
          required: true
          description: MongoDB id
          schema:
            type: string      

      responses:
        201:
          description: Movie deleted succesfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Movie"

  


    

 